requirements.txt
discord.py
disputils
python-decouple
youtube_dl
pynacl
